:package: azure ad integration: Refactor azure code, now it looks cleaner
