+++ affair/api/root/handler.py
@@ -1,0 +1,2 @@
+async def root():
+    return {"message": "Hello World"}
