+++ affair/azure.py
@@ -1,61 +1,49 @@
 
+import os
 import uvicorn
 
 from fastapi import FastAPI, Security
 from fastapi.middleware.cors import CORSMiddleware
-from pydantic import AnyHttpUrl, BaseSettings, Field
-from typing import Union
-from fastapi_azure_auth import SingleTenantAzureAuthorizationCodeBearer
 
-class Settings(BaseSettings):
-    SECRET_KEY: str = Field('my super secret key', env='SECRET_KEY')
-    BACKEND_CORS_ORIGINS: list[Union[str, AnyHttpUrl]] = ['http://localhost:8000']
-    OPENAPI_CLIENT_ID: str = Field(default='', env='OPENAPI_CLIENT_ID')
-    APP_CLIENT_ID: str = Field(default='', env='APP_CLIENT_ID')
-    TENANT_ID: str = Field(default='', env='TENANT_ID')
+from api.root.handler import root
 
-    class Config:
-        env_file = '.env'
-        env_file_encoding = 'utf-8'
-        case_sensitive = True
+from services.azure.usecase.settings import Settings
+from services.azure.usecase.scheme import scheme
+from services.azure.usecase.load_config import load_config
 
-settings = Settings()
+from dotenv import load_dotenv
+
+load_dotenv()
+
+azure_settings = Settings()
 
 app = FastAPI(
-    swagger_ui_oauth2_redirect_url='/oauth2-redirect',
+    swagger_ui_oauth2_redirect_url=os.getenv("SWAGGER_UI_OAUTH2_REDIRECT_URL"),
     swagger_ui_init_oauth={
         'usePkceWithAuthorizationCodeGrant': True,
-        'clientId': settings.OPENAPI_CLIENT_ID,
+        'clientId': azure_settings.OPENAPI_CLIENT_ID,
     }
 )
 
-if settings.BACKEND_CORS_ORIGINS:
+# Handle CORS
+if azure_settings.BACKEND_CORS_ORIGINS:
     app.add_middleware(
         CORSMiddleware,
-        allow_origins=[str(origin) for origin in settings.BACKEND_CORS_ORIGINS],
+        allow_origins=[str(origin) for origin in azure_settings.BACKEND_CORS_ORIGINS],
         allow_credentials=True,
         allow_methods=['*'],
         allow_headers=['*'],
     )
 
-azure_scheme = SingleTenantAzureAuthorizationCodeBearer(
-    app_client_id=settings.APP_CLIENT_ID,
-    tenant_id=settings.TENANT_ID,
-    scopes={
-        f'api://{settings.APP_CLIENT_ID}/user_impersonation': 'user_impersonation',
-    },
+# Event handlers
+app.add_event_handler("startup", load_config)
+
+# Endpoints
+app.add_api_route(
+    path="/",
+    dependencies=[Security(scheme)],
+    endpoint=root
 )
 
-@app.on_event('startup')
-async def load_config() -> None:
-    """
-    Load OpenID config on startup.
-    """
-    await azure_scheme.openid_config.load_config()
-
-@app.get("/", dependencies=[Security(azure_scheme)])
-async def root():
-    return {"message": "Hello World"}
-
 if __name__ == "__main__":
-    uvicorn.run('azure:app', reload=True)
+    uvicorn.run('main:app', reload=True)
