+++ affair/graphql.py
@@ -1,6 +1,5 @@
 import graphene
 
-import uvicorn
 from fastapi import FastAPI
 
 from starlette.graphql import GraphQLApp
