+++ affair/services/azure/usecase/scheme.py
@@ -1,0 +1,12 @@
+from fastapi_azure_auth import SingleTenantAzureAuthorizationCodeBearer
+from services.azure.usecase.settings import Settings
+
+settings = Settings()
+
+scheme = SingleTenantAzureAuthorizationCodeBearer(
+    app_client_id=settings.APP_CLIENT_ID,
+    tenant_id=settings.TENANT_ID,
+    scopes={
+        f'api://{settings.APP_CLIENT_ID}/user_impersonation': 'user_impersonation',
+    },
+)
